requirements.txt
# Flask==3.1.0
# numpy==1.22.4
# pandas==2.0.3
# Pillow==11.0.0
# scikit_learn==1.3.2
# torch==2.2.2

# Flask
# numpy
# pandas
# Pillow
# scikit_learn
# torch
# open_clip_torch


Flask
numpy
pandas
Pillow
scikit-learn
torch
open-clip-torch
